+++ plugin/src/com/microsoft/alm/plugin/idea/tfvc/core/TFSCommittedChangesProvider.java
@@ -99,7 +99,7 @@ public class TFSCommittedChangesProvider implements CachingCommittedChangesProvi
         settings.USE_CHANGE_AFTER_FILTER = true;
         settings.USE_CHANGE_BEFORE_FILTER = true;
         settings.CHANGE_BEFORE = settings.CHANGE_AFTER = String.valueOf(((TfsRevisionNumber) number).getValue());
-        final FilePath filePath = VcsContextFactory.SERVICE.getInstance().createFilePathOn(file);
+        final FilePath filePath = VcsContextFactory.getInstance().createFilePathOn(file);
         final List<TFSChangeList> list = getCommittedChanges(settings, getLocationFor(filePath), 1);
         if (list.size() == 1) {
             return Pair.create(list.get(0), filePath);